move tests around a bit now that we have tests
