+++ README.rst
@@ -39,7 +39,7 @@ Writing:
    ...  'primary_title': 'Reference 43',
    ...  'abstract': 'Lorem ipsum'
    ...  }]
-   >>> filepath = 'tests/example_export_full.ris'
+   >>> filepath = 'export.ris'
    >>> with open(filepath, 'w') as bibliography_file:
    ...     rispy.dump(entries, bibliography_file)
